add helper for extracting resource info
