+++ app/src/pages/[resourceType]/search-result.tsx
@@ -3,6 +3,7 @@ import { Center, Divider, Paper, Text } from '@mantine/core';
 import { ArtifactResourceType, FhirArtifact, ResourceInfo } from '@/util/types/fhir';
 import ResourceCards from '@/components/ResourceCards';
 import { ExternalLink } from 'tabler-icons-react';
+import { extractResourceInfo } from '@/util/resourceCardUtils';
 
 export default function ResourceSearchResultsPage({
   resourceInfo,
@@ -78,16 +79,7 @@ export const getServerSideProps: GetServerSideProps<{
     const resources = bundle.entry;
     const resourceInfoArray = resources.reduce((acc: ResourceInfo[], entry) => {
       if (entry.resource && entry.resource.id) {
-        const identifier = entry.resource.identifier?.[0];
-        const resourceInfo: ResourceInfo = {
-          resourceType: checkedResourceType,
-          id: entry.resource.id,
-          identifier: identifier?.system && identifier?.value ? `${identifier.system}|${identifier.value}` : null,
-          name: entry.resource.name ?? null,
-          url: entry.resource.url ?? null,
-          version: entry.resource.version ?? null,
-          status: entry.resource.status ?? null
-        };
+        const resourceInfo = extractResourceInfo(entry.resource);
         acc.push(resourceInfo);
       }
       return acc;
