+++ app/src/pages/authoring/[resourceType]/index.tsx
@@ -4,6 +4,7 @@ import { useRouter } from 'next/router';
 import { ArtifactResourceType, FhirArtifact, ResourceInfo } from '@/util/types/fhir';
 import ResourceCards from '@/components/ResourceCards';
 import { Edit } from 'tabler-icons-react';
+import { extractResourceInfo } from '@/util/resourceCardUtils';
 
 export default function ResourceAuthoringPage() {
   const router = useRouter();
@@ -12,17 +13,7 @@ export default function ResourceAuthoringPage() {
   const resourceTypeQuery = trpc.draft.getDrafts.useQuery(resourceType as ArtifactResourceType);
 
   const resourceCardContent: ResourceInfo[] = (resourceTypeQuery.data as FhirArtifact[])?.map(resource => {
-    const identifier = resource.identifier?.[0];
-    const resourceInfo: ResourceInfo = {
-      resourceType: resourceType as ArtifactResourceType,
-      id: resource.id as string,
-      identifier: identifier?.system && identifier?.value ? `${identifier.system}|${identifier.value}` : null,
-      name: resource.name ?? null,
-      url: resource.url ?? null,
-      version: resource.version ?? null,
-      status: resource.status ?? null
-    };
-    return resourceInfo;
+    return extractResourceInfo(resource);
   });
 
   return (
@@ -34,14 +25,12 @@ export default function ResourceAuthoringPage() {
       </Center>
       <Divider my="md" />
       <div style={{ paddingTop: '18px' }}>
-        <Center>
-          <ResourceCards
-            resourceInfo={resourceCardContent}
-            resourceType={resourceType as ArtifactResourceType}
-            icon={<Edit size="24" />}
-            authoring={true}
-          />
-        </Center>
+        <ResourceCards
+          resourceInfo={resourceCardContent}
+          resourceType={resourceType as ArtifactResourceType}
+          icon={<Edit size="24" />}
+          authoring={true}
+        />
       </div>
     </div>
   );
